Convert IntroActivity.kt to Fragment
