+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/intro/IntroViewPagerAdapter.kt
@@ -7,9 +7,9 @@ import io.horizontalsystems.bankwallet.R
 
 class IntroViewPagerAdapter(fragmentManager: FragmentManager) : FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
     private val fragments = listOf<Fragment>(
-            IntroFragment.newInstance(R.string.Intro_Independence_Title, R.string.Intro_Independence_Description),
-            IntroFragment.newInstance(R.string.Intro_Knowledge_Title, R.string.Intro_Knowledge_Description),
-            IntroFragment.newInstance(R.string.Intro_Privacy_Title, R.string.Intro_Privacy_Description)
+            IntroSlideFragment.newInstance(R.string.Intro_Independence_Title, R.string.Intro_Independence_Description),
+            IntroSlideFragment.newInstance(R.string.Intro_Knowledge_Title, R.string.Intro_Knowledge_Description),
+            IntroSlideFragment.newInstance(R.string.Intro_Privacy_Title, R.string.Intro_Privacy_Description)
     )
 
     override fun getCount(): Int = fragments.size
